Deprecated teams webhook function
